xnova_c2/server/requirements.txt
rich
pyjwt
colorama
cryptography
websockets
toml
mnemonic
pyotp
flask
uuid
pycryptodome
